+++ experiments/standalone/Soil/richards_runoff.jl
@@ -83,7 +83,7 @@ end
 vg_α = SpaceVaryingInput(
     joinpath(
         soil_params_artifact_path,
-        "vGalpha_map_gupta_etal2020_2.5x2.5x4.nc",
+        "vGalpha_map_gupta_etal2020_1.0x1.0x4.nc",
     ),
     "α",
     subsurface_space;
@@ -101,7 +101,7 @@ function mask_vg_n(var, value)
     end
 end
 vg_n = SpaceVaryingInput(
-    joinpath(soil_params_artifact_path, "vGn_map_gupta_etal2020_2.5x2.5x4.nc"),
+    joinpath(soil_params_artifact_path, "vGn_map_gupta_etal2020_1.0x1.0x4.nc"),
     "n",
     subsurface_space;
     regridder_type,
@@ -115,7 +115,7 @@ hydrology_cm = vg_fields_to_hcm_field.(vg_α, vg_n)
 θ_r = SpaceVaryingInput(
     joinpath(
         soil_params_artifact_path,
-        "residual_map_gupta_etal2020_2.5x2.5x4.nc",
+        "residual_map_gupta_etal2020_1.0x1.0x4.nc",
     ),
     "θ_r",
     subsurface_space;
@@ -126,7 +126,7 @@ hydrology_cm = vg_fields_to_hcm_field.(vg_α, vg_n)
 ν = SpaceVaryingInput(
     joinpath(
         soil_params_artifact_path,
-        "porosity_map_gupta_etal2020_2.5x2.5x4.nc",
+        "porosity_map_gupta_etal2020_1.0x1.0x4.nc",
     ),
     "ν",
     subsurface_space;
@@ -135,7 +135,7 @@ hydrology_cm = vg_fields_to_hcm_field.(vg_α, vg_n)
 )
 ν .= mask_vg.(ν, 1.0)
 K_sat = SpaceVaryingInput(
-    joinpath(soil_params_artifact_path, "ksat_map_gupta_etal2020_2.5x2.5x4.nc"),
+    joinpath(soil_params_artifact_path, "ksat_map_gupta_etal2020_1.0x1.0x4.nc"),
     "Ksat",
     subsurface_space;
     regridder_type,